refactor(core): Rename policies to policy, remove unused policy.py
