+++ core/agent/policy/manager.py
@@ -2,8 +2,8 @@ from __future__ import annotations
 
 import logging
 
-from core.agent.policies.discovery import discover_and_register
-from core.agent.policies.registry import CALLABLE_REGISTRY
+from core.agent.policy.discovery import discover_and_register
+from core.agent.policy.registry import CALLABLE_REGISTRY
 from app.utils.sqlite_client import SQLiteClient
 from app.policies.store import PolicyStore
 from app.schema.pydantic_models import EventType
